Cleanup and fix compile warnings

+++ ui/src/main/kotlin/bz/stew/bracken/ui/util/Math.kt
@@ -15,7 +15,7 @@ class Math {
             var ti: Double = t / (d / 2)
             if (t < 1) return c / 2 * ti * ti + b
             ti--
-            return -(c.toDouble()) / 2 * (ti * (t - 2) - 1) + b
+            return -(c) / 2 * (ti * (t - 2) - 1) + b
         }
 
     }
